Move signing decorators to correct package
