+++ services/distribution/src/main/java/app/coronawarn/server/services/distribution/assembly/diagnosiskeys/util/DateTime.java
@@ -48,8 +48,7 @@ public class DateTime {
    * Returns a set of all {@link LocalDateTime hours} that are associated with the submission timestamps of the
    * specified {@link DiagnosisKey diagnosis keys} and the specified {@link LocalDate date}.
    */
-  public static Set<LocalDateTime> getHours(LocalDate currentDate,
-      Collection<DiagnosisKey> diagnosisKeys) {
+  public static Set<LocalDateTime> getHours(LocalDate currentDate, Collection<DiagnosisKey> diagnosisKeys) {
     return diagnosisKeys.stream()
         .map(DiagnosisKey::getSubmissionTimestamp)
         .map(DateTime::getLocalDateTimeFromHoursSinceEpoch)
